refactor: :fire: moved google drive modal to the related directory
