+++ packages/bot-web-ui/src/components/load-modal/google-drive/google-drive.tsx
@@ -7,10 +7,10 @@ import { useDBotStore } from 'Stores/useDBotStore';
 import {
     rudderStackSendGoogleDriveConnectEvent,
     rudderStackSendGoogleDriveDisconnectEvent,
-} from '../../../../analytics/rudderstack-common-events';
+} from '../../../analytics/rudderstack-common-events';
 import './google-drive.scss';
 
-const GoogleDrive = observer(() => {
+const GoogleDrive: React.FC = observer(() => {
     const { ui } = useStore();
     const { google_drive, load_modal } = useDBotStore();
     const { is_authorised, signIn, signOut } = google_drive;
