+++ packages/bot-web-ui/src/components/load-modal/load-modal.tsx
@@ -7,13 +7,13 @@ import { useDBotStore } from 'Stores/useDBotStore';
 import { rudderStackSendSwitchLoadStrategyTabEvent } from '../../analytics/rudderstack-bot-builder';
 import { rudderStackSendCloseEvent } from '../../analytics/rudderstack-common-events';
 import { LOAD_MODAL_TABS } from '../../analytics/utils';
-import GoogleDrive from '../../pages/dashboard/bot-list/google-drive';
+import GoogleDrive from './google-drive';
 import Local from './local';
 import LocalFooter from './local-footer';
 import Recent from './recent';
 import RecentFooter from './recent-footer';
 
-const LoadModal = observer(() => {
+const LoadModal: React.FC = observer(() => {
     const { ui } = useStore();
     const { load_modal, dashboard } = useDBotStore();
     const {
